requirements.txt
# =============================================================================
# Trading Bot — Python Dependencies
# Python 3.12+
# =============================================================================

# --- Core framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1

# --- Database ---
sqlalchemy>=2.0.30

# --- HTTP client (for alerting webhooks) ---
httpx>=0.27.0

# --- Configuration ---
pyyaml>=6.0.1

# --- Jinja2 templates for dashboard ---
jinja2>=3.1.4

# --- Charts (optional — for backtest reports) ---
matplotlib>=3.9.0

# --- Testing ---
pytest>=8.2.0
pytest-asyncio>=0.23.7

# --- Code quality ---
ruff>=0.4.0
